measuring requests times

+++ src/images.rs
@@ -1,7 +1,7 @@
 
 pub struct Images {
     pub srv_id: u16,
-    pub quads: u16,
+    pub quads: u64,
     pub obs_id: u16,
 }
